requirements.txt
altair==5.3.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
build==1.2.1
buildozer==1.5.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
detect==2020.12.3
dill==0.3.8
distlib==0.3.8
Django
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
huggingface-hub==0.25.1
idna==3.7
ipykernel==6.29.5
ipython
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx
numpy
onnx==1.16.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scipy
seaborn==0.13.2
selenium==4.25.0
setuptools==70.2.0
sh==1.14.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
stack-data==0.6.3
streamlit==1.36.0
sympy==1.12.1
tenacity==8.5.0
timeout-decorator==0.5.0
tokenizers==0.20.0
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.45.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.63
ultralytics-thop==2.0.0
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
